+++ packages/core/src/interpolation/interpolateUrl.ts
@@ -1,3 +1,5 @@
+import interpolateValue from "./interpolateValue";
+
 function slugify(text: string) {
   return text.trim().replaceAll(' ', '-').replaceAll('/', '-').toLowerCase()
 }
@@ -7,7 +9,7 @@ function interpolateUrl(url: string, values: Record<string, string>) {
     const value = values[key];
     const slugifiedValue = slugify(value);
 
-    url = url.replace(new RegExp(`{${key}}`, 'g'), slugifiedValue);
+    url = interpolateValue(url, key, slugifiedValue);
   });
 
   return url;