+++ packages/core/src/main.ts
@@ -3,11 +3,13 @@ export { default as detectLocaleFromCookie } from './detection/detectLocaleFromC
 export { default as detectLocaleFromNavigator } from './detection/detectLocaleFromNavigator'
 export { default as detectLocaleFromPathname } from './detection/detectLocaleFromPathname'
 
+export { default as interpolateValue } from './interpolation/interpolateValue'
+export { default as interpolateUrl } from './interpolation/interpolateUrl'
+
 export * from './format/decline'
 export * from './format/format'
 export { default as formatToSegments } from './format/formatToSegments'
 export { default as formatToString } from './format/formatToString'
-export { default as interpolateUrl } from './format/interpolateUrl'
 export { default as formatValue } from './format/formatValue'
 
 export { default as localizeMessage } from './localization/localizeMessage'
@@ -21,5 +23,4 @@ export * from './translation/utils'
 export * from './constants'
 export { default as psitta } from './psitta'
 export * from './utils'
-
 export * from './typings'