+++ packages/core/CHANGELOG.md
@@ -4,6 +4,20 @@ name: Changelog
 description: All notable changes to this project will be documented in this file.
 ---
 
+## [Unreleased]
+
+## Added
+
+- Added `interpolateValue()`
+
+## Changed
+
+- Renamed `formatValue()` to `format()`
+
+## Removed
+
+- Removed `interpolate()`
+
 ## [0.2.0] - 2024-03-31
 
 ## Added